kernel: BIG kernel syscall interface refactor
The new proc.Thread is the unit of execution in Telepath.
Each Thread belongs to one proc.Process. Process is just
a collection of capabilities (proc.Capability). In other
words a Process is a "namespace" that dictates what Threads
belonging to that Process can and cannot do. Consequently
Threads belonging to the same Process can automatically
share their address space which makes implementing certain
user space programs easier (for example the file system).